+++ client/src/game/temp.ts
@@ -14,13 +14,13 @@ import type { Floor } from "./models/floor";
 import type { ServerShape } from "./models/shapes";
 import type { BaseTemplate } from "./models/templates";
 import { addOperation } from "./operations/undo";
-import type { Shape } from "./shapes/shape";
+import type { IShape } from "./shapes/interfaces";
 import { applyTemplate } from "./shapes/templates";
 import { createShapeFromDict } from "./shapes/utils";
 import { Asset } from "./shapes/variants/asset";
 import { visionState } from "./vision/state";
 
-export function moveFloor(shapes: Shape[], newFloor: Floor, sync: boolean): void {
+export function moveFloor(shapes: IShape[], newFloor: Floor, sync: boolean): void {
     shapes = shapes.filter((s) => !s.isLocked);
     if (shapes.length === 0) return;
     const oldLayer = shapes[0].layer;
@@ -45,7 +45,7 @@ export function moveFloor(shapes: Shape[], newFloor: Floor, sync: boolean): void
     }
 }
 
-export function moveLayer(shapes: readonly Shape[], newLayer: Layer, sync: boolean): void {
+export function moveLayer(shapes: readonly IShape[], newLayer: Layer, sync: boolean): void {
     if (shapes.length === 0) return;
     const oldLayer = shapes[0].layer;
 
@@ -72,7 +72,7 @@ export function moveLayer(shapes: readonly Shape[], newLayer: Layer, sync: boole
     }
 }
 
-export function addShape(shape: ServerShape, sync: SyncMode): Shape | undefined {
+export function addShape(shape: ServerShape, sync: SyncMode): IShape | undefined {
     if (!floorStore.hasLayer(floorStore.getFloor({ name: shape.floor })!, shape.layer)) {
         console.log(`Shape with unknown layer ${shape.layer} could not be added`);
         return;