+++ CHANGELOG.md
@@ -13,6 +13,9 @@ These usually have no immediately visible impact on regular users
 -   Dice rolling tool
 -   Pointer to the draw tool
 -   Server Admin Commands
+-   Polygon edit tool
+    -   cut polygon in two / add extra node to polygon / remove node from polygon
+    -   accessible when selecting a polygon with the select tool in BUILD mode
 
 ### Changed
 
@@ -44,6 +47,7 @@ These usually have no immediately visible impact on regular users
     -   Fix startup not working correctly
     -   Fix unsnapped move of blocking shape not updating movement triangulation
 -   Composite shape tracker/aura toggles
+-   Movement block not updating directly when not using snapping
 
 ## [0.28.0] - 2021-07-21